More linter fixes

+++ examples/runners/ash/src/main.rs
@@ -392,7 +392,7 @@ impl RenderBase {
 
         let swapchain_loader = khr::Swapchain::new(&instance, &device);
 
-        let present_queue = unsafe { device.get_device_queue(queue_family_index as u32, 0) };
+        let present_queue = unsafe { device.get_device_queue(queue_family_index, 0) };
 
         let surface_format = {
             let acceptable_formats = {
@@ -1263,7 +1263,7 @@ unsafe extern "system" fn vulkan_debug_callback(
     _user_data: *mut std::os::raw::c_void,
 ) -> vk::Bool32 {
     let callback_data = *p_callback_data;
-    let message_id_number: i32 = callback_data.message_id_number as i32;
+    let message_id_number: i32 = callback_data.message_id_number;
 
     let message_id_name = if callback_data.p_message_id_name.is_null() {
         Cow::from("")
